Added the media player to the application.
Please note that as of this commit, you can resolve and run
the file 'com.github.axiopisty.plarpebu.bndrun' in the application
project.

When you do this, a small GUI will show appear that will provide
a media player with basic controls, load, play, pause, stop. Each
control is integrated with the media player backend.

As of this commit, using the Load button to load an MP3 file causes
an exception to be printed to the console stating:

  No line matching interface SourceDataLine supporting format PCM_SIGNED
  44100.0 Hz, 16 bit, stereo, 4 bytes/frame, little-endian is supported.

However, if you instead use the gogo shell to load the MP3 file, it
loads just fine and you can use the play, pause, and stop buttons in the
GUI to control playing the application. Here is an example how to
load an MP3 with the gogo command:

  mp3:load <canonical_path_to_mp3>

I haven't been able to figure out why loading an MP3 works from the
gogo shell, but not from within the GUI. I'm going to post this on
github to see if I can get some help.